Revert "Revert "Merge pull request #1727 from jprjr/fix-deprecated-warnings""
This reverts commit be9c1e74e0da75a3b98f614a38b2c744bd5040a8.

+++ playlists/storedplaylistspage.cpp
@@ -295,7 +295,7 @@ void StoredPlaylistsPage::removeInvalid()
 
         PlaylistsModel::PlaylistItem *pl=static_cast<PlaylistsModel::PlaylistItem *>(item);
         QList<quint32> toRemove;
-        for (quint32 i=0; i<pl->songs.count(); ++i) {
+        for (quint32 i=0; i<(quint32)pl->songs.count(); ++i) {
             const auto song = pl->songs.at(i);
             if (song->isInvalid()) {
                 toRemove.append(i);
